+++ tests/testthat/test-interaction-changes.R
@@ -2,7 +2,7 @@ RERUN <- FALSE
 
 skip_if(!check_dependencies(python = TRUE))
 
-init_env(quiet = TRUE)
+init_env()
 
 map <- readRDS("map.rds")